fix(ci): repair workshop and openapi validation
